gravel: rename 'local' to 'requirements'
